cherry pick from d3cb4135 Delete Branch - fix errors related to push request

+++ src/main/java/com/checkmarx/flow/service/FlowService.java
@@ -5,6 +5,7 @@ import com.checkmarx.flow.dto.*;
 import com.checkmarx.flow.dto.report.ScanReport;
 import com.checkmarx.flow.exception.ExitThrowable;
 import com.checkmarx.flow.exception.GitHubClientException;
+import com.checkmarx.flow.exception.GitHubClientRunTimeException;
 import com.checkmarx.flow.exception.MachinaException;
 import com.checkmarx.flow.utils.ScanUtils;
 import com.checkmarx.flow.utils.ZipUtils;
@@ -160,12 +161,17 @@ public class FlowService {
             }
 
 
-        }catch (CheckmarxException | GitAPIException e){
-            String extendedMessage = ExceptionUtils.getMessage(e);
-            log.error(extendedMessage, e);
-            Thread.currentThread().interrupt();
-            OperationResult scanCreationFailure = new OperationResult(OperationStatus.FAILURE, e.getMessage());
-            logRequest(request, scanId, cxFile, scanCreationFailure);
+            
+        }
+        catch (GitHubClientRunTimeException e){
+            //the error message is printed when the exception is thrown
+            //usually should occur during push event occuring on delete branch
+            //therefore need to eliminate the scan process but do not want to create 
+            //an error stuck trace in the log
+            return new ScanDetails(UNKNOWN_INT, UNKNOWN_INT, new CompletableFuture<ScanResults> (), false);
+        }
+        catch (CheckmarxException | GitAPIException e){
+            String extendedMessage = treatFailure(request, cxFile, scanId, e);
             throw new MachinaException("Checkmarx Error Occurred: " + extendedMessage);
         }
 
@@ -173,6 +179,15 @@ public class FlowService {
 
         this.scanDetails = new ScanDetails(projectId, scanId, osaScanId);
         return scanDetails;
+    }
+
+    private String treatFailure(ScanRequest request, File cxFile, Integer scanId, Exception e) {
+        String extendedMessage = ExceptionUtils.getMessage(e);
+        log.error(extendedMessage, e);
+        Thread.currentThread().interrupt();
+        OperationResult scanCreationFailure = new OperationResult(OperationStatus.FAILURE, e.getMessage());
+        logRequest(request, scanId, cxFile, scanCreationFailure);
+        return extendedMessage;
     }
 
 
@@ -246,7 +261,7 @@ public class FlowService {
             }
         }
         if(!projectExists || flowProperties.isAlwaysProfile()) {
-            log.info("Project is new, profiling source...");
+            log.info("Project doesn't exist, profiling source...");
             Sources sources = new Sources();
             switch (request.getRepoType()) {
                 case GITHUB:
@@ -265,6 +280,7 @@ public class FlowService {
                     log.warn("Profiling is not available for Azure DevOps");
                     break;
                 default:
+                    log.info("Nothing to profile");
                     break;
             }
             String preset = helperService.getPresetFromSources(sources);
@@ -614,10 +630,12 @@ public class FlowService {
             String ownerId = determineTeamAndOwnerID(request);
             
             String projectName = determineProjectName(request);
-
+            
             Integer projectId = determinePresetAndProjectId(request, ownerId, projectName);
 
-            cxService.deleteProject(projectId);
+            if(projectId != UNKNOWN_INT) {
+                cxService.deleteProject(projectId);
+            }
 
         } catch (CheckmarxException | MachinaException e) {
             log.error("Error delete branch " + e.getMessage());
